fix: :bug: Refactor to custom hook
